# docs/api/openapi.json
{
  "components": {
    "schemas": {
...
      "CambiarPinRequest": {
        "properties": {
          "pinActual": {
            "pattern": "^\\d{6}$",
            "type": "string"
          },
          "pinNuevo": {
            "pattern": "^\\d{6}$",
            "type": "string"
          }
        },
        "required": [
          "pinActual",
          "pinNuevo"
        ],
        "type": "object"
      },
...
      "ComercioActivoRequest": {
        "properties": {
          "comercioId": {
            "format": "uuid",
            "type": "string"
          }
        },
        "required": [
          "comercioId"
        ],
        "type": "object"
      },
      "ComercioActivoResponse": {
        "properties": {
          "comercio": {
            "$ref": "#/components/schemas/EspacioResponse"
          },
          "permisos": {
            "example": [
              "consumptions.register"
            ],
            "items": {
              "type": "string"
            },
            "type": "array"
          }
        },
        "required": [
          "comercio",
          "permisos"
        ],
        "type": "object"
      },
      "CorreoResponse": {
        "properties": {
          "correo": {
            "example": "marta@lavecina.co",
            "type": "string"
          }
        },
        "required": [
          "correo"
        ],
        "type": "object"
      },
      "CorreoYContrasenaRequest": {
        "properties": {
          "contrasena": {
            "example": "almuerzo2026",
            "minLength": 8,
            "type": "string"
          },
          "correo": {
            "example": "marta@lavecina.co",
            "type": "string"
          },
          "pinActual": {
            "description": "PIN actual, para confirmar que eres tú",
            "pattern": "^\\d{6}$",
            "type": "string"
          }
        },
        "required": [
          "correo",
          "contrasena",
          "pinActual"
        ],
        "type": "object"
      },
...
      "DispositivoRequest": {
        "properties": {
          "id": {
            "description": "Id que el dispositivo generó para sí",
            "format": "uuid",
            "type": "string"
          },
          "modelo": {
            "example": "Moto E13",
            "type": "string"
          },
          "plataforma": {
            "enum": [
              "ANDROID",
              "IOS",
              "WEB"
            ],
            "type": "string"
          },
          "versionApp": {
            "example": "0.2.0",
            "type": "string"
          },
          "versionSo": {
            "example": "Android 13",
            "type": "string"
          }
        },
        "required": [
          "id",
          "plataforma"
        ],
        "type": "object"
      },
...
      "EspacioResponse": {
        "properties": {
          "comercioId": {
            "format": "uuid",
            "type": "string"
          },
          "invitacionPendiente": {
            "description": "Lo invitaron como cajero y aún no ha entrado a este negocio",
            "type": "boolean"
          },
          "nombre": {
            "example": "Restaurante La Vecina",
            "type": "string"
          },
          "roles": {
            "description": "OWNER, CASHIER o CUSTOMER",
            "example": [
              "CASHIER"
            ],
            "items": {
              "type": "string"
            },
            "type": "array"
          },
          "tipoNegocio": {
            "description": "Código del tipo de negocio",
            "example": "RESTAURANT",
            "type": "string"
          }
        },
        "required": [
          "comercioId",
          "nombre",
          "tipoNegocio",
          "roles",
          "invitacionPendiente"
        ],
        "type": "object"
      },
...
      "IngresoConContrasenaRequest": {
        "properties": {
          "contrasena": {
            "example": "almuerzo2026",
            "type": "string"
          },
          "correo": {
            "example": "marta@lavecina.co",
            "type": "string"
          },
          "dispositivo": {
            "$ref": "#/components/schemas/DispositivoRequest"
          }
        },
        "required": [
          "correo",
          "contrasena",
          "dispositivo"
        ],
        "type": "object"
      },
      "IngresoConPinRequest": {
        "properties": {
          "celular": {
            "description": "Celular como lo escribe la persona",
            "example": "310 000 0102",
            "type": "string"
          },
          "dispositivo": {
            "$ref": "#/components/schemas/DispositivoRequest"
          },
          "pin": {
            "example": "482915",
            "pattern": "^\\d{6}$",
            "type": "string"
          }
        },
        "required": [
          "celular",
          "pin",
          "dispositivo"
        ],
        "type": "object"
      },
      "IngresoResponse": {
        "properties": {
          "nombre": {
            "example": "Jhon",
            "type": "string"
          },
          "requiereCambioDePin": {
            "description": "Entró con un PIN temporal: debe crear el suyo antes de seguir",
            "type": "boolean"
          },
          "sesion": {
            "$ref": "#/components/schemas/SesionResponse"
          },
          "tokenCambio": {
            "description": "Solo si requiereCambioDePin. Dura 10 minutos.",
            "type": "string"
          }
        },
        "required": [
          "requiereCambioDePin",
          "nombre"
        ],
        "type": "object"
      },
...
      "PinNuevoRequest": {
        "properties": {
          "dispositivo": {
            "$ref": "#/components/schemas/DispositivoRequest"
          },
          "pinNuevo": {
            "example": "730284",
            "pattern": "^\\d{6}$",
            "type": "string"
          },
          "tokenCambio": {
            "description": "Token que entregó el ingreso con PIN temporal",
            "type": "string"
          }
        },
        "required": [
          "tokenCambio",
          "pinNuevo",
          "dispositivo"
        ],
        "type": "object"
      },
...
      "RenovarSesionRequest": {
        "properties": {
          "tokenRenovacion": {
            "type": "string"
          }
        },
        "required": [
          "tokenRenovacion"
        ],
        "type": "object"
      },
      "RespuestaErrorDto": {
        "properties": {
          "codigo": {
            "example": "HORARIO_SE_CRUZA",
            "type": "string"
          },
          "message": {
            "example": "Ese horario se cruza con otro del mismo día en esta sede.",
            "type": "string"
          },
          "statusCode": {
            "example": 422,
            "type": "number"
          }
        },
        "required": [
          "statusCode",
          "codigo",
          "message"
        ],
        "type": "object"
      },
...
      "SesionResponse": {
        "properties": {
          "espacios": {
            "items": {
              "$ref": "#/components/schemas/EspacioResponse"
            },
            "type": "array"
          },
          "segundosAcceso": {
            "example": 900,
            "type": "number"
          },
          "tokenAcceso": {
            "description": "Va en Authorization: Bearer. Dura 15 minutos.",
            "type": "string"
          },
          "tokenRenovacion": {
            "description": "Se cambia por uno nuevo en cada renovación; guárdelo seguro.",
            "type": "string"
          },
          "usuario": {
            "$ref": "#/components/schemas/UsuarioResponse"
          }
        },
        "required": [
          "tokenAcceso",
          "tokenRenovacion",
          "segundosAcceso",
          "usuario",
          "espacios"
        ],
        "type": "object"
      },
      "UsuarioResponse": {
        "properties": {
          "id": {
            "format": "uuid",
            "type": "string"
          },
          "nombre": {
            "example": "Jhon",
            "type": "string"
          }
        },
        "required": [
          "id",
          "nombre"
        ],
        "type": "object"
      }
    },
    "securitySchemes": {
      "bearer": {
        "bearerFormat": "JWT",
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "contact": {},
    "description": "El vecino aliado de los negocios del Putumayo. Contrato para la app y el panel.",
    "title": "API de VECI",
    "version": "0.1.0"
  },
  "openapi": "3.0.0",
  "paths": {
    "/cuenta/comercio-activo": {
      "post": {
        "operationId": "elegirComercio",
        "parameters": [],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ComercioActivoRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ComercioActivoResponse"
                }
              }
            },
            "description": ""
          },
          "401": {
            "description": "Sin sesión o sesión cerrada"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RespuestaErrorDto"
                }
              }
            },
            "description": "No es uno de mis negocios"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Elige el negocio con el que trabajo",
        "tags": [
          "Mi cuenta"
        ]
      }
    },
    "/cuenta/correo-y-contrasena": {
      "put": {
        "operationId": "definirCorreoYContrasena",
        "parameters": [],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CorreoYContrasenaRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CorreoResponse"
                }
              }
            },
            "description": ""
          },
          "401": {
            "description": "Sin sesión o sesión cerrada"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RespuestaErrorDto"
                }
              }
            },
            "description": "El correo es de otra cuenta"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RespuestaErrorDto"
                }
              }
            },
            "description": "Contraseña débil"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Correo y contraseña para entrar al panel (HU-02-02)",
        "tags": [
          "Mi cuenta"
        ]
      }
    },
    "/cuenta/espacios": {
      "get": {
        "operationId": "listarMisEspacios",
        "parameters": [],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/EspacioResponse"
                  },
                  "type": "array"
                }
              }
            },
            "description": ""
          },
          "401": {
            "description": "Sin sesión o sesión cerrada"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Comercios donde trabajo o soy cliente",
        "tags": [
          "Mi cuenta"
        ]
      }
    },
    "/cuenta/pin": {
      "put": {
        "operationId": "cambiarMiPin",
        "parameters": [],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CambiarPinRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "204": {
            "description": ""
          },
          "401": {
            "description": "Sin sesión o sesión cerrada"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RespuestaErrorDto"
                }
              }
            },
            "description": "PIN débil"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Cambia mi PIN",
        "tags": [
          "Mi cuenta"
        ]
      }
    },
...
    "/sesion/con-contrasena": {
      "post": {
        "operationId": "entrarConContrasena",
        "parameters": [],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/IngresoConContrasenaRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IngresoResponse"
                }
              }
            },
            "description": ""
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RespuestaErrorDto"
                }
              }
            },
            "description": "No coinciden o bloqueada"
          }
        },
        "summary": "Entra al panel con correo (HU-02-02)",
        "tags": [
          "Sesión"
        ]
      }
    },
    "/sesion/con-pin": {
      "post": {
        "operationId": "entrarConPin",
        "parameters": [],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/IngresoConPinRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IngresoResponse"
                }
              }
            },
            "description": ""
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RespuestaErrorDto"
                }
              }
            },
            "description": "No coinciden o bloqueada"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RespuestaErrorDto"
                }
              }
            },
            "description": "Cuenta pausada por VECI"
          }
        },
        "summary": "Entra con celular y PIN (HU-02-01)",
        "tags": [
          "Sesión"
        ]
      }
    },
    "/sesion/pin-nuevo": {
      "post": {
        "operationId": "crearPinNuevo",
        "parameters": [],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PinNuevoRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SesionResponse"
                }
              }
            },
            "description": ""
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RespuestaErrorDto"
                }
              }
            },
            "description": "El paso se venció"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RespuestaErrorDto"
                }
              }
            },
            "description": "PIN débil"
          }
        },
        "summary": "Crea el PIN propio tras uno temporal",
        "tags": [
          "Sesión"
        ]
      }
    },
    "/sesion/renovar": {
      "post": {
        "operationId": "renovarSesion",
        "parameters": [],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RenovarSesionRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SesionResponse"
                }
              }
            },
            "description": ""
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RespuestaErrorDto"
                }
              }
            },
            "description": "Sesión cerrada o vencida"
          }
        },
        "summary": "Cambia el token de renovación por otro",
        "tags": [
          "Sesión"
        ]
      }
    },
    "/sesion/salir": {
      "post": {
        "operationId": "salir",
        "parameters": [],
        "responses": {
          "204": {
            "description": ""
          },
          "401": {
            "description": "Sin sesión o sesión cerrada"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Cierra la sesión en este dispositivo",
        "tags": [
          "Sesión"
        ]
      }
    },
...
  },
  "servers": [],
...
}